move element click test to element_click folder, split element click test, fix tests.
