+++ webdriver/tests/element_click/support/close_window.html
@@ -1,0 +1,9 @@
+<html>
+    <body onload="checkload()">
+        <script>
+        function checkload() {
+            window.close()
+        }
+        </script>
+    </body>
+</html>
