+++ webdriver/tests/element_click/support/input.html
@@ -1,0 +1,3 @@
+<html>
+    <input type=text value="Hello World">
+</html>
